Reorganize middleware in seperate files
