+++ server/validators/index.js
@@ -1,0 +1,6 @@
+const validationMiddleWareFactory = require("./middleware");
+const prizeSchema = require("./schemas/prizes");
+
+module.exports = {
+  validatePrizesQueries: validationMiddleWareFactory(prizeSchema),
+};
